move engine test to test_util
