+++ src/metric-engine/src/engine.rs
@@ -274,50 +274,3 @@ impl MetricEngineInner {
         data_region_request
     }
 }
-
-#[cfg(test)]
-mod test {
-    use std::time::Duration;
-
-    use common_telemetry::info;
-
-    use super::*;
-    use crate::test_util::TestEnv;
-
-    #[tokio::test]
-    async fn create_metadata_region() {
-        common_telemetry::init_default_ut_logging();
-
-        let env = TestEnv::new().await;
-        let mito = env.mito();
-        let engine = MetricEngine {
-            inner: Arc::new(MetricEngineInner { mito }),
-        };
-        let engine_dir = env.data_home();
-        let region_dir = join_dir(&engine_dir, "test_metric_region");
-
-        let region_id = RegionId::new(1, 2);
-        let region_create_request = RegionCreateRequest {
-            engine: METRIC_ENGINE_NAME.to_string(),
-            column_metadatas: vec![],
-            primary_key: vec![],
-            options: HashMap::new(),
-            region_dir: "test_metric_region".to_string(),
-        };
-
-        // create the region
-        engine
-            .handle_request(region_id, RegionRequest::Create(region_create_request))
-            .await
-            .unwrap();
-
-        // assert metadata region's dir
-        let metadata_region_dir = join_dir(&region_dir, METADATA_REGION_SUBDIR);
-        let exist = tokio::fs::try_exists(region_dir).await.unwrap();
-        assert!(exist);
-
-        // check mito engine
-        let metadata_region_id = utils::to_metadata_region_id(region_id);
-        let result = env.mito().get_metadata(metadata_region_id).await.unwrap();
-    }
-}
